Implemented all changes requested.
